services/data-ingestion/requirements.txt
websockets==12.0
google-cloud-pubsub==2.19.0
